refactor(패키지 분리) : 각 클래스 패키지 분리
 - domain, util, service, view 패키지로 분리
