--- src/main/java/empty.txt
@@ -1,6 +1,0 @@
-1. 입력한 금액에 해당하는 로또를 발급해준다.
-2. 자동을 선택한 경우 구매한 개수만큼의 1~45 범위의 6자리 숫자를 발급해준다.
-3. 당첨번호 1~45범위의 6자리 숫자와 1개의 보너스 번호를 랜덤하게 생성한다.
-4. 구매한 번호와 당첨번호를 비교하여 당첨 복권의 수와 금액을 계산한다.
-5. 당첨된 복권에 대한 통계와 수익률을 출력한다.
-
