+++ src/main/java/lotto/Main.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.service.LottoManager;
+
 public class Main {
 
     public static void main(String[] args) {
