+++ src/main/java/lotto/domain/Lotto.java
@@ -1,9 +1,7 @@
-package lotto;
+package lotto.domain;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
+
 public class Lotto {
 
     private List<Integer> lotto;
@@ -35,8 +33,6 @@ public class Lotto {
             str.append(lotto.get(i)+", ");
         }
         str.append(lotto.get(lotto.size()-1));
-
-//        str.append(lotto.stream().map(integer ->toString()).collect(Collectors.joining(", ")));
         str.append("]\n");
         return str.toString();
     }
