+++ src/main/java/lotto/domain/LottoGenerator.java
@@ -1,4 +1,7 @@
-package lotto;
+package lotto.domain;
+
+import lotto.domain.Lotto;
+import lotto.util.RandomUtil;
 
 import java.util.ArrayList;
 import java.util.Collections;
