+++ src/main/java/lotto/util/RandomUtil.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.util;
 
 import java.util.Random;
