+++ src/test/java/lotto/domain/LottoReviewTest.java
@@ -1,5 +1,9 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.Lotto;
+import lotto.domain.LottoReview;
+import lotto.domain.Lottos;
+import lotto.domain.WonLotto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -9,7 +13,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class LottoReviewTest {
 
     private WonLotto wonLotto;
-    private  LottoReview lottoReview;
+    private LottoReview lottoReview;
     @BeforeEach
     void setUp() {
         wonLotto = new WonLotto(new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6)), 7);
