+++ src/test/java/lotto/domain/LottosTest.java
@@ -1,5 +1,9 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.Lotto;
+import lotto.domain.LottoRank;
+import lotto.domain.Lottos;
+import lotto.domain.WonLotto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
