+++ src/test/java/lotto/domain/WonLottoTest.java
@@ -1,5 +1,8 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.Lotto;
+import lotto.domain.LottoRank;
+import lotto.domain.WonLotto;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
