+++ src/test/java/lotto/util/RandomUtilTest.java
@@ -1,6 +1,6 @@
-package lotto;
+package lotto.util;
 
-import org.junit.jupiter.api.Assertions;
+import lotto.util.RandomUtil;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -12,7 +12,7 @@ public class RandomUtilTest {
     @DisplayName("랜덤값의 범위가 1~45이내인지 확인")
     public void testRandomValueRange(){
 
-        int randomValue=RandomUtil.getRandomValue();
+        int randomValue= RandomUtil.getRandomValue();
         assertThat(randomValue).isBetween(1,45);
     }
